Bounding Light Volumes + Improvements and Bug Fixes
A first implemention of spherical bounding volumes for all lights. This improves performance especially on scenes with a lot of point lights. Lit/Unlit geometry rendering is not needed anymore, but pixel light information is stored into the gbuffer. Lighting is moved to a separate lighting pass. Reworked the mesh classes to aid inheritance. Lightbuffer structures and manager is used for instanced light rendering. Added log levels + more logging messages to further trace errors. An issue related to glfence syncing was fixed (i'm still not sure how tf that worked...) that caused indefinite hanging of the app.