[WIP] Partial Commit
In the process of migrating the project to NET 5 and opentk 4.0. Unfortunately this causes serious issues with the UI. Therefore I decided to re-implement the UI using ImGui. ImGUI_SDL_ModelViewer will be the new project that will host that functionality. I've also commited some fixes on the main app, but since I'm also updating the MVCore codebase chances are there will be issues.