Namespace changes + TransformationSystem Implementation
Merged MVCore.GMDL to MVCore. I think there is no reason to have so many namespaces. The engine should have a few very solid namespaces so that it can be as compact as possible to other applications.

First implementation of the transformation system. Reworked entity/transformData/transformController/transformComponent classes. The main idea is that an entity that has transform should have a transformComponent. The register function of the TransformationSystem takes responsibility of generating the TransformData, TransformController (if needed) as well as the wrapper TransformComponent and also adds it to the entity. The TransformController class is introduced so that it can be used by the animation and physics systems as well. It should provide a means to pass transformation states to any object that should be transformed.