Repaired Camera System
Smooth camera movement. Reverted back to pitch yaw system. I also tested incremental quaternion calculations ended up to the same result but the thing is that  its more expensive and in general not worth it. TODO, check if I can use quaternions to take advantage of the per frame interpolation that can be achieved by the transform controller.