Made subprojects compilable. Removed resource system
Moved resources to their corresponding system. For the majority of rendering resources, several managers are introduced that belong to the rendering system and are primarily used from that system only.

Pretty happy with the code structure so far, 

Shaders fail to compile due to errors in shader file processing and resolving.